+++ packages/varlet-ui-playground/auto-imports.d.ts
@@ -2,6 +2,7 @@
 export {}
 declare global {
   const Dialog: typeof import('@varlet/ui')['Dialog']
+  const Locale: typeof import('@varlet/ui')['Locale']
   const Snackbar: typeof import('@varlet/ui')['Snackbar']
   const StyleProvider: typeof import('@varlet/ui')['StyleProvider']
 }