+++ scripts/build.mjs
@@ -4,11 +4,14 @@ import { resolve } from 'path'
 
 const CWD = process.cwd()
 const PKG_CLI = resolve(CWD, './packages/varlet-cli')
+const PKG_VITE_PLUGINS = resolve(CWD, './packages/varlet-vite-plugins')
 const PKG_ICONS = resolve(CWD, './packages/varlet-icons')
 const PKG_UI = resolve(CWD, './packages/varlet-ui')
 const PKG_SHARED = resolve(CWD, './packages/varlet-shared')
 
 export const buildCli = () => execa('pnpm', ['build'], { cwd: PKG_CLI })
+
+export const buildVitePlugins = () => execa('pnpm', ['build'], { cwd: PKG_VITE_PLUGINS })
 
 export const buildShared = () => execa('pnpm', ['build'], { cwd: PKG_SHARED })